[#22] QueryDSL 적용

See merge request s11-final/S11P31A501!21

+++ backend/jaegwan/src/main/java/com/bwmanager/jaegwan/ingredient/dto/IngredientDetailResponse.java
@@ -15,5 +15,5 @@ public class IngredientDetailResponse {
 
     private LocalDateTime purchaseDate;
     private double amount;
-    private long leftExpirationDay;
+    private int leftExpirationDay;
 }
